Huge structural reorganization to ensure word answer object does not reset
